+++ src/rdbg/configurations/debugConfiguration.ts
@@ -6,3 +6,17 @@ export interface DebugConfiguration extends vscode.DebugConfiguration {
   skipPaths: string[];
   type: DebugType;
 }
+
+export interface SocketDebugConfiguration extends DebugConfiguration {
+  host?: never;
+  port?: never;
+  socket: string;
+  socketTimeoutMs?: number;
+}
+
+export interface TcpDebugConfiguration extends DebugConfiguration {
+  host: string;
+  port: number;
+  socket?: never;
+  socketTimeoutMs?: never;
+}